requirements.txt
fastapi
uvicorn
jinja2
python-multipart
openai-whisper
torch
g4f